Final changes

+++ Movies.Application/Respository/IAsyncRepository.cs
@@ -9,7 +9,7 @@ namespace Movies.Application.Respository
     public interface IAsyncRepository<T> where T : class
     {
         Task<IEnumerable<T>> GetAllAsync();
-        Task<T> GetAsync(int id);
+        Task<T> GetAsync(ushort id);
         Task<T> AddAsync(T entity);
         Task DeleteAsync(T entity);
         Task UpdateAsync(T entity);
